# src/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIRIS EPM API
  description: |
    AIRIS Enterprise Performance Management (EPM) API provides comprehensive monitoring, 
    analytics, and performance optimization capabilities for enterprise systems.
    
    ## Features
    - Real-time performance monitoring
    - Data quality validation
    - Advanced analytics and reporting
    - Caching and optimization
    - Alert and notification system
    
    ## Authentication
    API uses Bearer token authentication. Include your API key in the Authorization header.
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per API key.
    
  version: "1.0.0"
  contact:
    name: AIRIS EPM Support
    url: https://airis-epm.com/support
    email: support@airis-epm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.airis-epm.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /performance/cache/{type}/{key}:
    get:
      tags:
        - Performance
      summary: Get cached data
      description: Retrieve data from cache
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          description: Cache type identifier
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Cache key
      responses:
        '200':
          description: Cached data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheDataResponse'
        '404':
          description: Cache data not found
    
    post:
      tags:
        - Performance
      summary: Store data in cache
      description: Store data in cache with specified key
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          description: Cache type identifier
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Cache key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CacheDataRequest'
      responses:
        '201':
          description: Data cached successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

    delete:
      tags:
        - Performance
      summary: Invalidate cache
      description: Remove data from cache
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          description: Cache type identifier
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Cache key
      responses:
        '200':
          description: Cache invalidated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Data Quality APIs
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: Get alerts
      description: Retrieve system alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, acknowledged]
          description: Alert status filter
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Alert severity filter
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of alerts to return
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'

    post:
      tags:
        - Alerts
      summary: Create alert rule
      description: Create a new alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleRequest'
      responses:
        '201':
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CacheDataResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        cached:
          type: boolean
        ttl:
          type: integer
    CacheDataRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: object
        ttl:
          type: integer
        strategy:
          type: string
          enum: [lru, lfu, fifo]
...
    AlertsResponse:
      type: object
      properties:
        success:
          type: boolean
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        total_count:
          type: integer
    Alert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, resolved, acknowledged]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        rule_id:
          type: string
...
    AlertRuleRequest:
      type: object
      required:
        - name
        - condition
        - severity
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
          default: true
        notification_channels:
          type: array
          items:
            type: string
    AlertRuleResponse:
      type: object
      properties:
        success:
          type: boolean
        rule:
          $ref: '#/components/schemas/AlertRule'
    AlertRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        condition:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...